feat: better route protection on client
