+++ src/context/index.ts
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const authContext = createContext<any>(null);
